backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
